// src/pages/modules/website-unit/profile/history/data/types.ts
export interface IHistoryUnit {
  id_unit_sejarah: string
  id_satuan_organisasi: string
  gambar_url: string
  gambar_key: string
  isi_sejarah: string
  tahun: string
  urutan: number
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_user_created: string
  nama_user_updated: string
}